Remove the word "Example" in the file and class names.
